refactor: ECS redesign — attributes, helpers, delete obsolete types
- Replace GetRequiredSystems()/GetProvidedComponents() virtual overrides with
  [RequiresSystem] and [Provides] attributes on bridge classes
- Add Get<T>(), Set<T>(), Add<T>(), Has<T>() helper methods on EcsComponentBridge
- Bridges now serialize component structs directly as [SerializeField] (A3 pattern)
- Component structs use [NonSerialized] on runtime fields (config stays visible)
- Delete IBridge, IEntityContext, IComponentData, ComponentData<T>, EntityTemplate
- Delete EcsConfigAsset SO — MaxEntities config moved to EcsSystemRunner
- EcsRuntime is now a simple static cache set by EcsSystemRunner
- Remove CreateEntity(IBridge) overload from EcsWorld
- Bridge methods changed from public to internal (PushToEcs, PullFromEcs etc)
- Update all game bridges: DodgeBridge, MoveEntityBridge, AnimationBridge, PlayerInputBridge

Library types reduced from 22 to 17. Zero metadata overrides per bridge.

+++ Assets/HelloDev/Entities/Runtime/Core/IEcsSystem.cs
@@ -12,6 +12,6 @@ namespace HelloDev.Entities
         void Execute(EcsWorld world, List<int> entities, float deltaTime);
         void Dispose();
         // Used by the runner to filter entities. Implement via EcsSystemBase.RequiredComponents.
-        int GetRequiredMask();
+        long GetRequiredMask();
     }
 }
